Switched frontend server from Caddy to Nginx
Due to the fact that RGV will be running behind an edge proxy that
terminates SSL, Nginx will be preferred over Caddy due to its higher
performance with static files.

Since Caddy will not have to be built and compressed on every
deployment, this will also speed up the build pipeline somewhat.